requirements.txt
numpy>=1.24.0
pandas>=2.0.0
opencv-python>=4.8.0
pyaudio>=0.2.13
scipy>=1.11.0
h5py>=3.9.0
pyarrow>=13.0.0
ffmpeg-python>=0.2.0
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.40.0
datasets>=2.18.0
huggingface-hub>=0.22.0
openai-whisper>=20231117
ultralytics>=8.0.0
fastapi>=0.110.0
uvicorn>=0.27.0
pybullet>=3.2.5
pyyaml>=6.0
tqdm>=4.66.0
psutil>=5.9.0
feather-format>=0.4.1
pgvector>=0.2.4
sqlalchemy>=2.0.0
typer>=0.9.0
rich>=13.0.0
pytest>=7.4.0
python-multipart>=0.0.9
